+++ RELEASE-NOTES.md
@@ -52,6 +52,7 @@
 * More clear error messages in `snow snowpark build` command
 * Adding support for any source supported by `pip` in `snow snowpark`.
 * Fixed version parsing for packages lookup on Snowflake Anaconda Channel
+* Fix handling database/schema/role identifiers containing dashes
 
 # v2.1.2